backend/requirements.txt
# Django Core
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
mysqlclient==2.2.1

# Authentication
djangorestframework-simplejwt==5.3.1
django-allauth==0.57.0
dj-rest-auth==5.0.2

# Google OAuth
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Celery
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Environment
python-decouple==3.8
python-dotenv==1.0.0

# WSGI Server
gunicorn==21.2.0

# Utilities
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3.post1

# Development
django-debug-toolbar==4.2.0

